move kef_to_ph5 back to test_ph5tostationxml.py, remove redundant code in test_base.py
