+++ ph5/clients/tests/test_ph5tostationxml.py
@@ -8,9 +8,46 @@ from mock import patch
 
 from testfixtures import OutputCapture
 
+from ph5.utilities import kef2ph5
 from ph5.clients import ph5tostationxml
-from ph5.core.tests.test_base import LogTestCase, TempDirTestCase, \
-     kef_to_ph5
+from ph5.core.tests.test_base import LogTestCase, TempDirTestCase,\
+    initialize_ex
+
+
+def kef_to_ph5(ph5path, nickname, kefpath, keflist, ex=None):
+    """
+    Add kef to ph5file or to experiment (if ex is passed).
+    (The task of deleting table before adding the table should happen before
+    calling this function. If it is required to have a delete function for all,
+    it should be written in nuke_table.py)
+
+    :para ph5path: path to ph5 file
+    :type ph5path: string
+    :para kefpath: path to kef files
+    :type kefpath: string
+    :para keflist: A list of kef file names
+    :type keflist: list of string
+    :para ex: ph5 experiment from caller
+    :para ex: ph5 experiment object
+    :result: the tables in the kef files will be added to ph5 file or the
+    reference ex (if passed)
+    """
+
+    if ex is None:
+        with OutputCapture():
+            kef2ph5.EX = initialize_ex(nickname, ph5path, True)
+    else:
+        kef2ph5.EX = ex
+
+    kef2ph5.PH5 = os.path.join(ph5path, nickname)
+    kef2ph5.TRACE = False
+
+    for kef in keflist:
+        kef2ph5.KEFFILE = os.path.join(kefpath, kef)
+        kef2ph5.populateTables()
+
+    if ex is None:
+        kef2ph5.EX.ph5close()
 
 
 class TestPH5toStationXMLParser_main(LogTestCase, TempDirTestCase):
@@ -27,9 +64,10 @@ class TestPH5toStationXMLParser_main(LogTestCase, TempDirTestCase):
             with OutputCapture() as out:
                 ph5tostationxml.main()
                 output = out.captured.strip().split("\n")
-        self.assertEqual(
-            output[1],
-            "AA|PH5 TEST SET|2019-06-29T18:08:33|2019-09-28T14:29:39|1")
+                self.assertEqual(
+                    output[1],
+                    "AA|PH5 TEST SET|2019-06-29T18:08:33|"
+                    "2019-09-28T14:29:39|1")
 
 
 class TestPH5toStationXMLParser(LogTestCase, TempDirTestCase):
